Refactor the way tests are written in Java cores. (#15)
Instead of consuming inputs slice by slice, a slice being made of exactly one value for each and every input slot, this allows to specify the whole set of input values for a given input slot.
The main benefit is that it allows to build nodes that will consume more than one input value to compute the final value. An example would be for a node to find the min/max of all values passed through a single input slot.

Because of the change, the scala core was broken. Because this was just a toy language that I don't find interesting, the whole core has been dropped.